pyproject.toml
[tool.poetry]
name = "ytpodcast"
version = "0.1.0"
description = ""
authors = ["Carlo De Pieri <depieri.carlo@gmail.com>"]

packages = [
    { include = "ytpodcast" }
]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.78.0"
rfeed = "^1.1.1"
vcrpy = "^4.1.1"
pytube = "^12.1.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
